+++ src/main/java/com/shinhan/knockknock/domain/dto/user/TokenResponse.java
@@ -1,6 +1,7 @@
-package com.shinhan.knockknock.domain.dto;
+package com.shinhan.knockknock.domain.dto.user;
 
 import com.shinhan.knockknock.domain.entity.UserRoleEnum;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,12 +11,12 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class LoginUserResponse {
-    private Long userNo;
-    private String userId;
-    private String userPassword;
-    private String userName;
+public class TokenResponse {
+    @Schema(example = "1")
+    private long userNo;
+    @Schema(example = "PROTECTOR")
     private UserRoleEnum userType;
-    private String userSimplePassword;
+    @Schema(example = "accessToken")
+    private String accessToken;
     private String message;
 }
